Added start of documentation
I have created a file on overleaf which contains the start of the documentation for the project. I've begun with MCMC samplers and a brief explanation of RWM. This is by no means complete, and I will significantly revise it in future. Please let me know if/when you want access to the document and I can then share it with you on overleaf